Appointment Plugin Complete

+++ plugin-appointments/src/com/skillstorm/services/AppointmentService.java
@@ -4,7 +4,6 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -71,8 +70,9 @@ public class AppointmentService {
 						result.getString("title"),
 						result.getString("description"),
 						result.getString("datetime"),
-						result.getString("ownerId"),
-						result.getString("attendeeId")));
+						result.getString("organizerId"),
+						result.getString("attendeeId"),
+						result.getBoolean("checkedIn")));
 			}
 			return appointments;
 		} catch(Exception e) {
@@ -101,12 +101,95 @@ public class AppointmentService {
 						result.getString("title"),
 						result.getString("description"),
 						result.getString("datetime"),
-						result.getString("ownerId"),
-						result.getString("attendeeId"));
+						result.getString("organizerId"),
+						result.getString("attendeeId"),
+						result.getBoolean("checkedIn"));
 				return resultDept;
 			}
 			else {
 				throw new GeneralException("Could not find a Appointment with " + col + " " + value + ".");
+			}
+		} catch(Exception e) {
+			throw new GeneralException(e);
+		} finally {
+			IOUtil.closeQuietly(stmt);
+			IOUtil.closeQuietly(conn);
+		}
+	}
+	
+	public List<Appointment> getAppointmentsByOrganizer(String organizerId) throws GeneralException {
+		Connection conn = null;
+		PreparedStatement stmt = null;
+		try {
+			ResultSet result = executeQuery(conn, stmt, "SELECT * FROM " + dbName + " WHERE organizerId = \"" + organizerId +"\"");
+			List<Appointment> appointments = new ArrayList<>();
+			while(result.next()) {
+				appointments.add(new Appointment(result.getInt("id"),
+						result.getString("title"),
+						result.getString("description"),
+						result.getString("datetime"),
+						result.getString("organizerId"),
+						result.getString("attendeeId"),
+						result.getBoolean("checkedIn")));
+			}
+			return appointments;
+		} catch(Exception e) {
+			throw new GeneralException(e);
+		} finally {
+			IOUtil.closeQuietly(stmt);
+			IOUtil.closeQuietly(conn);
+		}
+	}
+	
+	public List<Appointment> getAppointmentsByAttendee(String attendeeId) throws GeneralException {
+		Connection conn = null;
+		PreparedStatement stmt = null;
+		try {
+			ResultSet result = executeQuery(conn, stmt, "SELECT * FROM " + dbName + " WHERE attendeeId = \"" + attendeeId + "\"");
+			List<Appointment> appointments = new ArrayList<>();
+			while(result.next()) {
+				appointments.add(new Appointment(result.getInt("id"),
+						result.getString("title"),
+						result.getString("description"),
+						result.getString("datetime"),
+						result.getString("organizerId"),
+						result.getString("attendeeId"),
+						result.getBoolean("checkedIn")));
+			}
+			return appointments;
+		} catch(Exception e) {
+			throw new GeneralException(e);
+		} finally {
+			IOUtil.closeQuietly(stmt);
+			IOUtil.closeQuietly(conn);
+		}
+	}
+	
+	/**
+	 * getDepartmentByColumn()
+	 * Modularization of get one function.
+	 * @param column
+	 * @param value
+	 * @return
+	 * @throws GeneralException
+	 */
+	public Appointment getAppointmentById(String id) throws GeneralException {
+		Connection conn = null;
+		PreparedStatement stmt = null;
+		try {
+			ResultSet result = executeQuery(conn, stmt, "SELECT * FROM " + dbName + " WHERE id = " + id);
+			if(result.next()) {
+				Appointment resultDept = new Appointment(result.getInt("id"),
+						result.getString("title"),
+						result.getString("description"),
+						result.getString("datetime"),
+						result.getString("organizerId"),
+						result.getString("attendeeId"),
+						result.getBoolean("checkedIn"));
+				return resultDept;
+			}
+			else {
+				throw new GeneralException("Could not find a Appointment with id " + id);
 			}
 		} catch(Exception e) {
 			throw new GeneralException(e);
@@ -123,13 +206,13 @@ public class AppointmentService {
 	 * @return
 	 * @throws GeneralException
 	 */
-	public Appointment createAppointment(String title, String description, String datetime, String ownerId, String attendeeId) throws GeneralException {
+	public Appointment createAppointment(String title, String description, String datetime, String organizerId, String attendeeId) throws GeneralException {
 		Connection conn = null;
 		PreparedStatement stmt = null;
 		try {
 			//	Create new Department
-			executeUpdate(conn, stmt, "INSERT INTO " + dbName + " (title, description, datetime, ownerId, attendeeId) VALUES"
-					+ " (\"" + title + "\", \"" + description + "\", \"" + datetime + "\", \"" + ownerId + "\", \"" + attendeeId + "\")");
+			executeUpdate(conn, stmt, "INSERT INTO " + dbName + " (title, description, datetime, organizerId, attendeeId) VALUES"
+					+ " (\"" + title + "\", \"" + description + "\", \"" + datetime + "\", \"" + organizerId + "\", \"" + attendeeId + "\")");
 			
 			//	Return newly created Department
 			return getAppointmentByColumn("title", title);
@@ -150,17 +233,37 @@ public class AppointmentService {
 	 * @return
 	 * @throws GeneralException
 	 */
-	public Appointment updateAppointmentByColumn(String col, String value, String id) throws GeneralException {
+	public Appointment updateAppointmentById(String id, String title, String description, String datetime, String organizerId, String attendeeId) throws GeneralException {
 		Connection conn = null;
 		PreparedStatement stmt = null;
 		try {
 			//	Update existing department
-			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET " + col + " = " + value + " WHERE id = " + id);
+			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET title = \"" + title + "\" WHERE id = " + id);
+			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET description = \"" + description + "\" WHERE id = " + id);
+			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET datetime = \"" + datetime + "\" WHERE id = " + id);
+			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET organizerId = \"" + organizerId + "\" WHERE id = " + id);
+			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET attendeeId = \"" + attendeeId + "\" WHERE id = " + id);
 
 			//	Return newly updated Department
-			return getAppointmentByColumn("id", id);
+			return getAppointmentById(id);
 		} catch(Exception e) {
-			throw new GeneralException("UPDATE " + dbName + " SET " + col + " = " + value + " WHERE id = " + id + e);
+			throw new GeneralException(e);
+		} finally {
+			IOUtil.closeQuietly(stmt);
+			IOUtil.closeQuietly(conn);
+		}
+	}
+	
+	public Appointment appointmentCheckIn(String id) throws GeneralException {
+		Connection conn = null;
+		PreparedStatement stmt = null;
+		try {
+			executeUpdate(conn, stmt, "UPDATE " + dbName + " SET checkedIn = 1 WHERE id = " + id);
+
+			//	Return newly updated Department
+			return getAppointmentById(id);
+		} catch(Exception e) {
+			throw new GeneralException(e);
 		} finally {
 			IOUtil.closeQuietly(stmt);
 			IOUtil.closeQuietly(conn);
@@ -179,9 +282,15 @@ public class AppointmentService {
 		Connection conn = null;
 		PreparedStatement stmt = null;
 		try {
-			executeUpdate(conn, stmt, "DELETE FROM " + dbName + " WHERE id = " + id);
 			
-			return getAppointmentByColumn("id", id);
+			Appointment deleted = getAppointmentById(id);
+			if(deleted == null) {
+				throw new GeneralException("Could not find a Department with id " + id);
+			}
+			else {
+				executeUpdate(conn, stmt, "DELETE FROM " + dbName + " WHERE id = " + id);
+				return deleted;
+			}
 		} catch(Exception e) {
 			throw new GeneralException(e);
 		} finally {
